refactor shell commands.
The shell commands now provide their own result processing functions.

Fix bug in that short/long range scans where not returned as result
to the ship command, but sent as a sync event on the side.

This did not play out well for the klingons,
that look at the data programmatically..